requirements.txt
django-cms>=3.7,<3.8
djangocms-admin-style>=1.4,<1.5
django-treebeard>=4.0,<5.0

djangocms-text-ckeditor>=3.7,<3.9
djangocms-link>=2.5,<2.6
djangocms-style>=2.2,<2.3
djangocms-googlemap>=1.3,<1.4
djangocms-snippet>=2.2,<2.3
djangocms-video>=2.1,<2.3
djangocms-file>=2.3,<3.0
djangocms-picture>=2.3,<2.4
djangocms-bootstrap4>=1.5,<1.6
easy_thumbnails
django-filer>=1.3
Django<2.0
django-classy-tags>=0.9
django-sekizai>=1.0
django-mptt>0.9
django-formtools>=2.1,<2.2
html5lib>=1.0.1
Pillow>=3.0
six
pytz